third_party/blink/web_tests/wpt_internal/script_tools/form_removal_submit_crash.https.html

<!DOCTYPE html>
<title>Form removal on submit crash test</title>
<link rel="author" href="mailto:dom@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../webmcp/resources/webmcp-helpers.js"></script>

<form toolname="crash_test" tooldescription="the crashy form!" toolautosubmit target="iframe" action="about:blank">
  <input name="dummy" type="text">
</form>
<iframe name="iframe"></iframe>

<script>
promise_test(async t => {
  assert_implements(navigator.modelContextTesting,
      "modelContextTesting should be available");

  const form = document.querySelector('form');
  let submitFired = false;

  form.addEventListener('submit', (e) => {
    submitFired = true;
    e.preventDefault();
    e.respondWith(Promise.resolve());

    // This synchronously unregisters the tool. The previously crashy path was
    // when the internal form infrastructure relied on the form still being
    // registered as a tool immediately after the submit event is fired.
    // Specifically, we reacted to the Promise above, and the Promise reaction
    // relied on the tool registration still being valid.
    //
    // See https://crbug.com/493850055.
    form.remove();
  });

  await waitForTool('crash_test');

  // Execute the tool. This synchronously fires the `submit` event at the form.
  // The tool execution to reject since the tool definition is updated/cancelled
  // when the form is removed from the document.
  const promise = navigator.modelContextTesting.executeTool('crash_test', '{}');
  promise_rejects_dom(t, "UnknownError", promise,
      "Form execution fails due to mid-execution unregistration.");

  assert_true(submitFired);
}, "Form removal on submit with respondWith does not crash the renderer");
</script>
